# backend/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: VITBON Cloud API
  version: 1.0.0
  description: REST API для мобильной кассы VITBON

servers:
  - url: https://api.vitbon.ru/api/v1
    description: Production
  - url: https://staging.vitbon.ru/api/v1
    description: Staging

paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Авторизация кассира
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin, deviceId]
              properties:
                pin:
                  type: string
                  minLength: 4
                  maxLength: 6
                deviceId:
                  type: string
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  cashier:
                    type: object
                    properties:
                      id: { type: string }
                      name: { type: string }
                      role: { type: string, enum: [ADMIN, SENIOR_CASHIER, CASHIER] }
        '401':
          description: Неверный ПИН

  /auth/logout:
    post:
      tags: [Auth]
      summary: Выход кассира
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Сессия отозвана
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /audit/sync:
    post:
      tags: [Audit]
      summary: Загрузить buffered audit events с Android
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events:
                  type: array
                  items:
                    type: object
                    required: [id, action, timestamp]
                    properties:
                      id: { type: string, format: uuid }
                      cashierId: { type: string, format: uuid, nullable: true }
                      deviceId: { type: string, nullable: true }
                      action: { type: string }
                      details: { type: string, nullable: true }
                      timestamp: { type: integer, format: int64 }
      responses:
        '200':
          description: Buffered audit events ingested into backend audit trail
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed: { type: integer }
                  failed:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        error: { type: string }
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /checks:
    post:
      tags: [Checks]
      summary: Создать чек
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckDto'
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Созданный чек
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckDto'
    get:
      tags: [Checks]
      summary: Список чеков
      security:
        - BearerAuth: []
      parameters:
        - name: shiftId
          in: query
          schema: { type: string }
        - name: date
          in: query
          schema: { type: string }
        - name: since
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Список чеков
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckDto'

  /checks/{id}:
    get:
      tags: [Checks]
      summary: Получить чек по id
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Чек
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckDto'
        '404':
          description: Чек не найден

  /checks/sync:
    post:
      tags: [Checks]
      summary: Синхронизация чеков (push)
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                checks:
                  type: array
                  items:
                    $ref: '#/components/schemas/CheckDto'
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Результат обработки
          content:
            application/json:
              schema:
                type: object
                properties:
                  processed: { type: integer }
                  failed:
                    type: array
                    items:
                      type: object
                      properties:
                        localUuid: { type: string }
                        error: { type: string }

  /products:
    get:
      tags: [Products]
      summary: Каталог товаров (delta sync)
      security:
        - BearerAuth: []
      parameters:
        - name: since
          in: query
          schema:
            type: integer
            format: int64
          description: Только товары обновлённые после этого timestamp (ms)
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Delta-обновление каталога
          content:
            application/json:
              schema:
                type: object
                properties:
                  products: { type: array, items: { $ref: '#/components/schemas/ProductDto' } }
                  deletedIds: { type: array, items: { type: string } }
                  serverTimestamp: { type: integer, format: int64 }

  /products/sync:
    post:
      tags: [Products]
      summary: Синхронизация каталога товаров (compat echo)
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                products: { type: array, items: { $ref: '#/components/schemas/ProductDto' } }
                deletedIds: { type: array, items: { type: string } }
                serverTimestamp: { type: integer, format: int64 }
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Echo payload
          content:
            application/json:
              schema:
                type: object
                properties:
                  products: { type: array, items: { $ref: '#/components/schemas/ProductDto' } }
                  deletedIds: { type: array, items: { type: string } }
                  serverTimestamp: { type: integer, format: int64 }

  /shifts/{cashierId}:
    get:
      tags: [Shifts]
      summary: Смены кассира
      security:
        - BearerAuth: []
      parameters:
        - name: cashierId
          in: path
          required: true
          schema: { type: string }
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Список смен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShiftDto'

  /statuses:
    get:
      tags: [System]
      summary: Статусы системы
      security:
        - BearerAuth: []
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Текущие статусы
          content:
            application/json:
              schema:
                type: object
                properties:
                  ofdQueueLength: { type: integer }
                  lastSyncTimestamp: { type: integer, format: int64, description: 0 when the backend has no device-scoped history yet }
                  cloudServerOk: { type: boolean }
                  licenseStatus: { type: string, enum: [ACTIVE, EXPIRED, GRACE_PERIOD, UNLICENSED] }

  /license/check:
    post:
      tags: [Licensing]
      summary: Проверка статуса лицензии
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId: { type: string }
      responses:
        '200':
          description: Статус лицензии
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [ACTIVE, EXPIRED, GRACE_PERIOD, UNLICENSED] }
                  expiryDate: { type: integer, format: int64, nullable: true }
                  graceUntil: { type: integer, format: int64, nullable: true }

  /reports:
    get:
      tags: [Reports]
      summary: Отчёт по продажам
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          required: true
          schema: { type: string }
        - name: shiftId
          in: query
          schema: { type: string }
        - name: since
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Отчёт по продажам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesReportDto'

  /reports/sales:
    get:
      tags: [Reports]
      summary: Отчёт по продажам (compat alias)
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          required: true
          schema: { type: string }
        - name: shiftId
          in: query
          schema: { type: string }
        - name: since
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Отчёт по продажам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesReportDto'

  /reports/movement:
    get:
      tags: [Reports]
      summary: Отчёт по движению товаров
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          required: true
          schema: { type: string }
        - name: since
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200':
          description: Отчёт по движению товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementReportDto'

  /documents/acceptance:
    post:
      tags: [Documents]
      summary: Приёмка товара
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string, enum: [ACCEPTANCE] }
                items: { type: array, items: { $ref: '#/components/schemas/DocumentItemDto' } }
                timestamp: { type: integer, format: int64 }
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200': { description: Принято }

  /documents/writeoff:
    post:
      tags: [Documents]
      summary: Списание товара
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200': { description: Принято }

  /documents/inventory:
    post:
      tags: [Documents]
      summary: Инвентаризация
      security:
        - BearerAuth: []
      responses:
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '200': { description: Принято }

  /egais/incoming:
    post:
      tags: [ЕГАИС]
      summary: Прокси приёмки накладной ЕГАИС
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/xml:
            schema:
              type: string
              example: '<EgaisRequest>...</EgaisRequest>'
      responses:
        '200':
          description: XML-ответ upstream ЕГАИС, проксированный без преобразования.
          content:
            application/xml:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'

  /egais/status:
    get:
      tags: [ЕГАИС]
      summary: Безопасная проверка доступности маршрута ЕГАИС
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Backend-side reachability required for EGAIS routes.
          content:
            application/json:
              schema:
                type: object
                required: [available]
                properties:
                  available:
                    type: boolean
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /egais/tara:
    post:
      tags: [ЕГАИС]
      summary: Акт вскрытия тары
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/xml:
            schema:
              type: string
              example: '<ActChargeOnWrite>...</ActChargeOnWrite>'
      responses:
        '200':
          description: XML-ответ upstream ЕГАИС, проксированный без преобразования.
          content:
            application/xml:
              schema:
                type: string
        '409':
          description: Upstream ЕГАИС отклонил акт вскрытия тары.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'

  /chaseznak/validate:
    post:
      tags: [Честный ЗНАК]
      summary: Валидация DataMatrix-кода в системе маркировки
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        '200':
          description: Нормализованный результат проверки кода маркировки.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChaseznakValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'

  /chaseznak/sell:
    post:
      tags: [Честный ЗНАК]
      summary: Выбытие маркированного товара
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
                checkId: { type: string }
      responses:
        '200':
          description: Upstream-ответ на выбытие, проксированный без преобразования.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '202':
          description: Upstream принял выбытие в асинхронную обработку.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'

  /chaseznak/verify-age:
    post:
      tags: [Честный ЗНАК]
      summary: Верификация возраста (Цифровой ID Max)
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                qrData: { type: string }
      responses:
        '200':
          description: Результат верификации возраста.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgeVerificationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
      description: Opaque bearer token issued by /auth/login and validated against server-side session state.

  responses:
    UnauthorizedError:
      description: Unauthorized — missing bearer token or invalid/expired/revoked session.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              value:
                code: UNAUTHORIZED
                message: Missing bearer token
            invalidSession:
              value:
                code: UNAUTHORIZED
                message: Invalid session
    ForbiddenError:
      description: Forbidden — authenticated session exists, but role is not allowed for operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            roleForbidden:
              value:
                code: FORBIDDEN
                message: Forbidden
    ServiceUnavailableError:
      description: Service unavailable — integration endpoint is not configured or upstream integration is unavailable.
    NotImplementedError:
      description: Not implemented — endpoint is intentionally fail-closed until the production integration is configured.

  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, enum: [UNAUTHORIZED, FORBIDDEN] }
        message: { type: string }

    ChaseznakValidationResponse:
      type: object
      required: [barcode, status]
      properties:
        barcode: { type: string }
        status:
          type: string
          enum: [OK, NOT_IN_CIRCULATION, ALREADY_SOLD, EXPIRED, ERROR]
        productName: { type: string, nullable: true }
        expiryDate: { type: integer, format: int64, nullable: true }
        message: { type: string, nullable: true }

    AgeVerificationResponse:
      type: object
      required: [verified]
      properties:
        verified: { type: boolean }
        verificationId: { type: string, nullable: true }
        message: { type: string, nullable: true }

    CheckDto:
      type: object
      required: [id, type, total, createdAt]
      properties:
        id: { type: string }
        localUuid: { type: string }
        shiftId: { type: string, nullable: true }
        cashierId: { type: string, nullable: true }
        deviceId: { type: string }
        type:
          type: string
          enum: [SALE, RETURN, CORRECTION, CASH_IN, CASH_OUT]
        fiscalSign: { type: string, nullable: true }
        ffdVersion: { type: string }
        subtotal: { type: integer }
        discount: { type: integer }
        total: { type: integer }
        taxAmount: { type: integer }
        paymentType: { type: string }
        items: { type: array, items: { $ref: '#/components/schemas/CheckItemDto' } }
        createdAt: { type: integer, format: int64 }

    CheckItemDto:
      type: object
      properties:
        id: { type: string }
        productId: { type: string, nullable: true }
        barcode: { type: string, nullable: true }
        name: { type: string }
        quantity: { type: number }
        price: { type: integer }
        discount: { type: integer }
        vatRate: { type: string }
        total: { type: integer }

    ProductDto:
      type: object
      properties:
        id: { type: string }
        barcode: { type: string, nullable: true }
        name: { type: string }
        article: { type: string, nullable: true }
        price: { type: integer }
        vatRate: { type: string }
        categoryId: { type: string, nullable: true }
        stock: { type: number }
        egaisFlag: { type: boolean }
        chaseznakFlag: { type: boolean }
        updatedAt: { type: integer, format: int64 }

    ShiftDto:
      type: object
      properties:
        id: { type: string }
        cashierId: { type: string }
        deviceId: { type: string }
        openedAt: { type: integer, format: int64 }
        closedAt: { type: integer, format: int64, nullable: true }
        totalCash: { type: integer }
        totalCard: { type: integer }

    DocumentItemDto:
      type: object
      properties:
        productId: { type: string, nullable: true }
        barcode: { type: string, nullable: true }
        name: { type: string }
        quantity: { type: number }
        reason: { type: string, nullable: true }

    ProductSalesDto:
      type: object
      properties:
        name: { type: string }
        quantity: { type: number }
        total: { type: integer }

    SalesReportDto:
      type: object
      properties:
        totalChecks: { type: integer }
        returnChecks: { type: integer }
        totalRevenue: { type: integer }
        totalReturns: { type: integer }
        cashRevenue: { type: integer }
        cardRevenue: { type: integer }
        averageCheck: { type: integer }
        topProducts:
          type: array
          items:
            $ref: '#/components/schemas/ProductSalesDto'

    MovementReportItemDto:
      type: object
      properties:
        name: { type: string }
        income: { type: number }
        sales: { type: number }
        balance: { type: number }

    MovementReportDto:
      type: object
      properties:
        openingStock: { type: number }
        income: { type: number }
        sales: { type: number }
        returns: { type: number }
        writeoff: { type: number }
        closingStock: { type: number }
        items:
          type: array
          items:
            $ref: '#/components/schemas/MovementReportItemDto'